// models/survey.ts
export class Survey {
    SurveyId: number;
    Name: string;
    StartDate: string;
    EndDate: string;
    IsDeleted: boolean;
    CreatedBy: string;
    CreatedOn: string;
    ThemeId: number;
    SurveyStatusId: number;
    ClientId: number;
    IsTemplate: boolean;
    LogoName: string;
}